Fix component instantiation not parsed when only generic_map is there
instantiation as follows:
inst_m: my_module
generic map(
    MY_GENERIC => something
);
were seen as error because the grammar was forcing the port map to be
present. I changed it so the port map is also optional. However, I had
to set precedence on the concurrent_procedural_statement because of
conflict. To keep backward compatibility, I made sure the
concurrent_procedural_block was still the one taking priority in case of
    m_label: identifier();
It now parses these things as function call (due to VHDL-2008) which is
"fine" because it removes an error node which was my concern.
I also added a test to make sure the generic map only instance is
working as intended.